loop_interchange_1 fix stack overflow
The C++ original was using `std::array<std::array<float, N>, N>` (stack) for `Matrix`
and then putting that in `unique_ptr` (heap). I tried to make it as close as possible
with `Box<[[f32; N]; N]>`. That means we have to create the 400x400 array on the stack
before moving it to the heap.

That needs about a 4 MB stack. Rust (`cargo test`) runs it's tests in a
thread, which I think have a 2 MB stack by default on Linux. So boom.

I replaced the Boxed array with a vector of vectors, meaning we create
directly on the heap now. It should come out the same, but without the
stack overflow and less copying during setup.